Move testing of download virus scanning to a separate test module
