+++ ftw/file/tests/test_download.py
@@ -1,21 +1,13 @@
-from collective.clamav.interfaces import IAVScannerSettings
-from collective.clamav.testing import EICAR
-from ftw.builder import create
-from ftw.builder import Builder
 from ftw.file.interfaces import IFileDownloadedEvent
-from ftw.file.testing import FTW_FILE_AV_FUNCTIONAL_TESTING
 from ftw.file.testing import FTW_FILE_FUNCTIONAL_TESTING
 from ftw.testbrowser import browsing
 from ftw.testbrowser import LIB_TRAVERSAL
-from ftw.testbrowser.pages import statusmessages
 from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
-from plone.registry.interfaces import IRegistry
 from StringIO import StringIO
 from unittest import TestCase
 from webdav.common import rfc1123_date
 from zope.component import eventtesting
-from zope.component import getUtility
 from zope.component import queryMultiAdapter
 
 import AccessControl
@@ -127,32 +119,6 @@ class TestFileDownload(TestCase):
                          browser.headers['content-disposition'])
         self.assertEqual('bytes', browser.headers['accept-ranges'])
         self.assertEqual('application/msword', browser.headers['content-type'])
-
-
-class TestFileVirusScanning(TestCase):
-
-    layer = FTW_FILE_AV_FUNCTIONAL_TESTING
-
-    def setUp(self):
-        self.portal = self.layer['portal']
-        setRoles(self.portal, TEST_USER_ID, ['Manager'])
-        registry = getUtility(IRegistry)
-        registry.forInterface(IAVScannerSettings).clamav_enabled = False
-        transaction.commit()
-        self.context = create(Builder('file').attach_file_containing(
-                              EICAR, 'v1.txt'), processForm=False)
-        registry.forInterface(IAVScannerSettings).clamav_enabled = True
-        transaction.commit()
-
-    @browsing
-    def test_prevent_download_of_virus_file(self, browser):
-        """
-        Test we prevent the download of a file with virus (unless virus scanning is explicitly off)
-        """
-        browser.login().visit(self.context, view='@@download')
-        statusmessages.assert_message(
-            'Download blocked. The malware Eicar-Test-Signature has been found in the file.'
-        )
 
 
 class TestFileDownloadView(TestCase):
